chore: delete administrative-sdk
deleted the big sdk file which is no longer used
