+++ administrative-sdk/choiceRecognition.js
@@ -6,7 +6,7 @@
  max-len,
  no-unused-vars
  */
-'use strict';
+
 const autobahn = require('autobahn');
 const Student = require('../administrative-sdk/student').Student;
 const PronunciationAnalysis = require('../administrative-sdk/pronunciationAnalysis').PronunciationAnalysis;
